requirements.txt
numpy
scipy
py-pde
imageio